refactor(jpa): merge extension generation into existing DAO generators
Move search/remove extension function generation from separate
ZygardeJpaDaoExtensionGenerator/KspGenerator classes into the existing
ZygardeJpaDaoGenerator/KspGenerator. Extensions are still generated as
separate *DaoExtensions.kt files (required because Kotlin interface
default methods are not recognized as such by Spring Data 2.7.x
without -Xjvm-default=all).
