+++ modules-jpa/zygarde-jpa-codegen-ksp/src/main/kotlin/zygarde/codegen/ksp/generator/ZygardeJpaDaoKspGenerator.kt
@@ -9,13 +9,20 @@ import com.google.devtools.ksp.symbol.KSType
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
+import com.squareup.kotlinpoet.LambdaTypeName
+import com.squareup.kotlinpoet.MemberName
 import com.squareup.kotlinpoet.ParameterSpec
+import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeName
 import com.squareup.kotlinpoet.TypeSpec
+import com.squareup.kotlinpoet.UNIT
+import com.squareup.kotlinpoet.asClassName
 import com.squareup.kotlinpoet.ksp.toTypeName
 import com.squareup.kotlinpoet.ksp.writeTo
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.PageRequest
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.stereotype.Component
@@ -26,6 +33,12 @@ import zygarde.codegen.ksp.ZygardeJpaKspOptions.DAO_PACKAGE
 import zygarde.codegen.ksp.ZygardeJpaKspOptions.DAO_SUFFIX
 import zygarde.codegen.ksp.extension.generic
 import zygarde.codegen.ksp.extension.kotlin
+import zygarde.core.exception.BusinessException
+import zygarde.core.exception.ErrorCode
+import zygarde.data.api.PagingAndSortingRequest
+import zygarde.data.api.SortField
+import zygarde.data.jpa.search.EnhancedSearch
+import zygarde.data.jpa.search.SearchSpecBuilder
 
 class ZygardeJpaDaoKspGenerator(
   private val codeGenerator: CodeGenerator,
@@ -48,11 +61,15 @@ class ZygardeJpaDaoKspGenerator(
     val daoPackage = packageName(options.getOrDefault(DAO_PACKAGE, "data.dao"))
     val daoSuffix = options.getOrDefault(DAO_SUFFIX, "Dao")
 
+    val generateRemove = daoInherit?.contains("ZygardeEnhancedDao") == true
+    val toSpringDataSort = MemberName("zygarde.data.jpa.search.request", "toSpringDataSort")
+    val toSpringDataPageRequest = MemberName("zygarde.data.jpa.search.request", "toSpringDataPageRequest")
     elements.forEach { element ->
       val entityName = element.simpleName.asString()
       val daoName = "$entityName$daoSuffix"
       val entityTypeName = element.asType(emptyList()).toTypeName().copy(nullable = false)
       val idTypeName = element.findIdClass()
+      val daoType = ClassName(daoPackage, daoName)
 
       FileSpec.builder(daoPackage, daoName)
         .addType(
@@ -76,6 +93,18 @@ class ZygardeJpaDaoKspGenerator(
         )
         .build()
         .writeTo(codeGenerator, aggregating = false)
+
+      val searchContentType = searchContentLambdaType(entityTypeName)
+      buildExtensionFileSpec(
+        daoPackage,
+        daoName,
+        daoType,
+        entityTypeName,
+        searchContentType,
+        toSpringDataSort,
+        toSpringDataPageRequest,
+        generateRemove,
+      ).build().writeTo(codeGenerator, aggregating = false)
     }
 
     if (options.getOrDefault(DAO_COMBINE, "true") == "true") {
@@ -110,6 +139,120 @@ class ZygardeJpaDaoKspGenerator(
         .build()
         .writeTo(codeGenerator, aggregating = false)
     }
+  }
+
+  private fun buildExtensionFileSpec(
+    daoPackage: String,
+    daoName: String,
+    daoType: ClassName,
+    entityType: TypeName,
+    searchContentType: LambdaTypeName,
+    toSpringDataSort: MemberName,
+    toSpringDataPageRequest: MemberName,
+    generateRemove: Boolean,
+  ): FileSpec.Builder {
+    val fileSpec = FileSpec.builder(daoPackage, "${daoName}Extensions")
+      .addFunction(
+        FunSpec.builder("search")
+          .receiver(daoType)
+          .addParameter("searchContent", searchContentType)
+          .returns(List::class.asClassName().parameterizedBy(entityType))
+          .addStatement("return findAll(%T.buildSpec(searchContent))", SearchSpecBuilder::class)
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("search")
+          .receiver(daoType)
+          .addParameter(
+            "sorts",
+            List::class.asClassName().parameterizedBy(SortField::class.asClassName()).copy(nullable = true)
+          )
+          .addParameter("searchContent", searchContentType)
+          .returns(List::class.asClassName().parameterizedBy(entityType))
+          .addStatement(
+            "return sorts?.let { findAll(%T.buildSpec(searchContent), it.%M()) } ?: search(searchContent)",
+            SearchSpecBuilder::class,
+            toSpringDataSort,
+          )
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("search")
+          .receiver(daoType)
+          .addParameter("searchContent", searchContentType)
+          .addParameter("limit", Int::class)
+          .returns(List::class.asClassName().parameterizedBy(entityType))
+          .addStatement(
+            "return findAll(%T.buildSpec(searchContent), %T.of(0, limit)).content",
+            SearchSpecBuilder::class,
+            PageRequest::class,
+          )
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("searchCount")
+          .receiver(daoType)
+          .addParameter("searchContent", searchContentType)
+          .returns(Long::class)
+          .addStatement("return count(%T.buildSpec(searchContent))", SearchSpecBuilder::class)
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("searchOne")
+          .receiver(daoType)
+          .addParameter("searchContent", searchContentType)
+          .returns(entityType.copy(nullable = true))
+          .addStatement(
+            "return findOne(%T.buildSpec(searchContent)).let { if (it.isPresent) it.get() else null }",
+            SearchSpecBuilder::class,
+          )
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("searchOneOrThrow")
+          .receiver(daoType)
+          .addParameter("errorCode", ErrorCode::class)
+          .addParameter("searchContent", searchContentType)
+          .returns(entityType)
+          .addStatement(
+            "return searchOne(searchContent) ?: throw %T(errorCode)",
+            BusinessException::class,
+          )
+          .build()
+      )
+      .addFunction(
+        FunSpec.builder("searchPage")
+          .receiver(daoType)
+          .addParameter("req", PagingAndSortingRequest::class)
+          .addParameter("searchContent", searchContentType)
+          .returns(Page::class.asClassName().parameterizedBy(entityType))
+          .addStatement(
+            "return findAll(%T.buildSpec(searchContent), req.%M())",
+            SearchSpecBuilder::class,
+            toSpringDataPageRequest,
+          )
+          .build()
+      )
+
+    if (generateRemove) {
+      fileSpec.addFunction(
+        FunSpec.builder("remove")
+          .receiver(daoType)
+          .addParameter("searchContent", searchContentType)
+          .returns(Int::class)
+          .addStatement("return delete(%T.buildSpec(searchContent))", SearchSpecBuilder::class)
+          .build()
+      )
+    }
+
+    return fileSpec
+  }
+
+  private fun searchContentLambdaType(entityType: TypeName): LambdaTypeName {
+    return LambdaTypeName.get(
+      receiver = EnhancedSearch::class.asClassName().parameterizedBy(entityType),
+      returnType = UNIT,
+    )
   }
 
   @OptIn(KspExperimental::class)
